requirements.txt
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
jupyter>=1.0.0
ipython>=7.30.0
pystan>=3.9.0
nest_asyncio>=1.6.0
